# OrderManagement.Api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Order Management API",
    "version": "1.0",
    "description": "Multi-tenant restaurant order management platform for Malaysia",
    "contact": {
      "name": "Order Management Support",
      "email": "support@ordermanagement.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5073",
      "description": "Local Development Server (HTTP)"
    },
    {
      "url": "https://localhost:7118",
      "description": "Local Development Server (HTTPS - requires SSL verification disabled)"
    }
  ],
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT Bearer token only"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@dinehub.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Admin@123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Admin", "Manager"]
          }
        }
      },
      "MenuItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "branchId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "name": {
            "type": "string",
            "example": "Dosa"
          },
          "category": {
            "type": "string",
            "example": "South Indian"
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "example": 12.50
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
          }
        }
      },
      "CreateMenuItemRequest": {
        "type": "object",
        "required": ["branchId", "name", "category", "price"],
        "properties": {
          "branchId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "name": {
            "type": "string",
            "example": "Nasi Lemak"
          },
          "category": {
            "type": "string",
            "example": "Malaysian"
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "example": 8.50
          },
          "isAvailable": {
            "type": "boolean",
            "example": true,
            "default": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
          }
        }
      },
      "UpdateMenuItemRequest": {
        "type": "object",
        "required": ["name", "category", "price", "isAvailable"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Dosa Special"
          },
          "category": {
            "type": "string",
            "example": "South Indian"
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "example": 15.00
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/v1/Auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticate user and receive JWT access token",
        "operationId": "Login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "admin@dinehub.com",
                "password": "Admin@123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                },
                "example": {
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1NTBlODQwMC1lMjliLTQxZDQtYTcxNi00NDY2NTU0NDAwMDAiLCJlbWFpbCI6ImFkbWluQGRpbmVodWIuY29tIiwidGVuYW50SWQiOiI1NTBlODQwMC1lMjliLTQxZDQtYTcxNi00NDY2NTU0NDAwMDIiLCJicmFuY2hJZCI6bnVsbCwicm9sZXMiOlsiQWRtaW4iLCJNYW5hZ2VyIl0sImV4cCI6MTczNDI3ODIyMiwiaXNzIjoiT3JkZXJNYW5hZ2VtZW50IiwiYXVkIjoiT3JkZXJNYW5hZ2VtZW50In0.example",
                  "roles": ["Admin", "Manager"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/Auth/seed-super-admin": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Seed super admin user",
        "description": "Create a super admin user (development only)",
        "operationId": "SeedSuperAdmin",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "superadmin@ordermanagement.local",
                "password": "SuperAdmin@123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Super admin created successfully"
          },
          "409": {
            "description": "Super admin already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/MenuItems": {
      "get": {
        "tags": ["Menu Items"],
        "summary": "Get all menu items",
        "description": "Retrieve all menu items for the current tenant/branch",
        "operationId": "GetMenuItems",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "branchId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "description": "Filter by branch ID",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          }
        ],
        "responses": {
          "200": {
            "description": "List of menu items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItemResponse"
                  }
                },
                "example": [
                  {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "branchId": "550e8400-e29b-41d4-a716-446655440001",
                    "name": "Dosa",
                    "category": "South Indian",
                    "price": 12.50,
                    "isAvailable": true,
                    "imageUrl": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
                  },
                  {
                    "id": "550e8400-e29b-41d4-a716-446655440003",
                    "branchId": "550e8400-e29b-41d4-a716-446655440001",
                    "name": "Nasi Lemak",
                    "category": "Malaysian",
                    "price": 8.50,
                    "isAvailable": true,
                    "imageUrl": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Menu Items"],
        "summary": "Create a new menu item",
        "description": "Create a new menu item for the current tenant/branch",
        "operationId": "CreateMenuItem",
        "security": [{"Bearer": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuItemRequest"
              },
              "example": {
                "branchId": "550e8400-e29b-41d4-a716-446655440001",
                "name": "Roti Canai",
                "category": "Malaysian",
                "price": 5.50,
                "isAvailable": true,
                "imageUrl": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Menu item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/MenuItems/{id}": {
      "get": {
        "tags": ["Menu Items"],
        "summary": "Get menu item by ID",
        "description": "Retrieve a specific menu item by its ID",
        "operationId": "GetMenuItem",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Menu item found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponse"
                },
                "example": {
                  "id": "550e8400-e29b-41d4-a716-446655440000",
                  "branchId": "550e8400-e29b-41d4-a716-446655440001",
                  "name": "Dosa",
                  "category": "South Indian",
                  "price": 12.50,
                  "isAvailable": true,
                  "imageUrl": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
                }
              }
            }
          },
          "404": {
            "description": "Menu item not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Menu Items"],
        "summary": "Update menu item",
        "description": "Update an existing menu item",
        "operationId": "UpdateMenuItem",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemRequest"
              },
              "example": {
                "name": "Dosa Special",
                "category": "South Indian",
                "price": 15.00,
                "isAvailable": true,
                "imageUrl": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Menu item updated successfully"
          },
          "404": {
            "description": "Menu item not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": ["Menu Items"],
        "summary": "Delete menu item",
        "description": "Delete a menu item",
        "operationId": "DeleteMenuItem",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "204": {
            "description": "Menu item deleted successfully"
          },
          "404": {
            "description": "Menu item not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/menu-management/items/{id}": {
      "get": {
        "tags": ["Menu Management"],
        "summary": "Get menu management item by ID",
        "description": "Get a specific navigation menu item",
        "operationId": "GetMenuManagementItem",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Menu management item found"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
...
}